+++ source/gear/riskymixture-midass.js
@@ -27,6 +27,6 @@ module.exports = new GearTemplate("Midas's Risky Mixture",
 		return generateModifierResultLines(combineModifierReceipts(receipts));
 	}
 ).setTargetingTags({ type: "single", team: "any" })
-	.setSidegrades("Potent Risky Mixture", "Thick Risky Mixture")
+	.setSidegrades("Potent Risky Mixture", "Unbreakable Risky Mixture")
 	.setModifiers({ name: "Poison", stacks: 4 }, { name: "Regen", stacks: 4 }, { name: "Curse of Midas", stacks: 1 })
 	.setDurability(15);