+++ source/gear/riskymixture-potent.js
@@ -25,6 +25,6 @@ module.exports = new GearTemplate("Potent Risky Mixture",
 		}
 	}
 ).setTargetingTags({ type: "single", team: "any" })
-	.setSidegrades("Midas's Risky Mixture", "Thick Risky Mixture")
+	.setSidegrades("Midas's Risky Mixture", "Unbreakable Risky Mixture")
 	.setModifiers({ name: "Poison", stacks: 6 }, { name: "Regen", stacks: 6 })
 	.setDurability(15);