+++ source/gear/riskymixture-unbreakable.js
@@ -1,8 +1,10 @@
 const { GearTemplate } = require('../classes');
 const { addModifier, changeStagger, generateModifierResultLines } = require('../util/combatantUtil');
+const { unbreakablePassive } = require('./descriptions/passives');
 
-module.exports = new GearTemplate("Thick Risky Mixture",
+module.exports = new GearTemplate("Unbreakable Risky Mixture",
 	[
+		unbreakablePassive,
 		["use", "Inflict @{mod0Stacks} @{mod0} on a target"],
 		["Critical💥", "Apply @{mod1} instead of @{mod0}"]
 	],
@@ -27,4 +29,4 @@ module.exports = new GearTemplate("Thick Risky Mixture",
 ).setTargetingTags({ type: "single", team: "any" })
 	.setSidegrades("Potent Risky Mixture", "Midas's Risky Mixture")
 	.setModifiers({ name: "Poison", stacks: 4 }, { name: "Regen", stacks: 4 })
-	.setDurability(30);
+	.setDurability(Infinity);